--- src/lib/vocab/account/accountRepo.ts
@@ -1,57 +1,0 @@
-import type {Result} from '@feltcoop/felt';
-import {Logger} from '@feltcoop/felt/util/log.js';
-import {blue, gray} from 'kleur/colors';
-
-import {PostgresRepo} from '$lib/db/PostgresRepo';
-import type {Account, AccountModel} from '$lib/vocab/account/account.js';
-import type {Database} from '$lib/db/Database';
-import type {ErrorResponse} from '$lib/util/error';
-import {toPasswordKey} from '$lib/util/password';
-
-const log = new Logger(gray('[') + blue('accountRepo') + gray(']'));
-
-export const accountRepo = (db: Database) =>
-	({
-		create: async (
-			name: string,
-			password: string,
-		): Promise<Result<{value: Account}, ErrorResponse>> => {
-			const passwordKey = await toPasswordKey(password);
-			const data = await db.sql<Account[]>`
-				INSERT INTO accounts (name, password) VALUES (
-					${name}, ${passwordKey}
-				) RETURNING *
-			`;
-			log.trace('created account', data[0]);
-			const account = data[0];
-			return {ok: true, value: account};
-		},
-		findById: async (
-			account_id: number,
-		): Promise<Result<{value: AccountModel}, {type: 'no_account_found'} & ErrorResponse>> => {
-			log.trace('loading account', account_id);
-			const data = await db.sql<AccountModel[]>`
-				SELECT account_id, name, created, updated FROM accounts WHERE account_id = ${account_id}
-			`;
-			if (data.length) {
-				log.trace('account found, returning', account_id);
-				return {ok: true, value: data[0]};
-			}
-			return {
-				ok: false,
-				type: 'no_account_found',
-				message: 'no account found',
-			};
-		},
-		findByName: async (
-			name: string,
-		): Promise<Result<{value: Account}, {type: 'no_account_found'} & ErrorResponse>> => {
-			const data = await db.sql<Account[]>`
-				SELECT account_id, name, password, created, updated FROM accounts WHERE name = ${name}
-			`;
-			if (data.length) {
-				return {ok: true, value: data[0]};
-			}
-			return {ok: false, type: 'no_account_found', message: 'no account found'};
-		},
-	} as const);
